Refactor MapProjectAccessError into own file
Simpilfy MapProject
