+++ react/src/components/MapProjectAccessError/MapProjectAccessError.module.css
@@ -1,0 +1,8 @@
+.errorContainer {
+    font-size: large;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    width: 100%;
+  }
