+++ react/src/pages/MapProject/MapProject.module.css
@@ -1,11 +1,3 @@
-.errorContainer {
-  font-size: large;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  width: 100%;
-}
 
 .root {
   display: flex;
