+++ react/src/pages/MapProject/MapProject.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
+import { useLocation, useParams } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
 
-import { Message, Button, LoadingSpinner } from '@tacc/core-components';
+import { LoadingSpinner } from '@tacc/core-components';
 
 import Map from '@hazmapper/components/Map';
 import AssetsPanel from '@hazmapper/components/AssetsPanel';
@@ -12,13 +12,12 @@ import { queryPanelKey, Panel } from '@hazmapper/utils/panels';
 import {
   useFeatures,
   useProject,
-  useAuthenticatedUser,
   useTileServers,
   useFeatureSelection,
   KEY_USE_FEATURES,
 } from '@hazmapper/hooks';
-import * as ROUTES from '@hazmapper/constants/routes';
 import MapProjectNavBar from '@hazmapper/components/MapProjectNavBar';
+import MapProjectAccessError from '@hazmapper/components/MapProjectAccessError';
 import MapControlBar from '@hazmapper/components/MapControlBar';
 import Filters from '@hazmapper/components/FiltersPanel/Filter';
 import { assetTypeOptions } from '@hazmapper/components/FiltersPanel/Filter';
@@ -28,62 +27,6 @@ import { MapPositionProvider } from '@hazmapper/context/MapContext';
 
 import styles from './MapProject.module.css';
 
-interface MapProjectAccessErrorProps {
-  error: any;
-}
-
-const MapProjectAccessError: React.FC<MapProjectAccessErrorProps> = ({
-  error,
-}) => {
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  const { data: authenticatedUser } = useAuthenticatedUser();
-  const isLoggedIn = !!authenticatedUser;
-
-  const getMessage = () => {
-    if (!error?.response) {
-      return 'Unable to load map project due to a server error';
-    }
-
-    switch (error.response.status) {
-      case 404:
-        return 'This map project does not exist';
-      case 403:
-        return isLoggedIn
-          ? "You don't have permission to access this map project"
-          : 'Please log in.'; /* no op as before this point, we ensure users are logged in non-public maps */
-      case 500:
-        return 'Unable to load map project due to a server error';
-      default:
-        return 'Unable to access this map';
-    }
-  };
-
-  const is403AndNotLoggedIn = error?.response?.status === 403 && !isLoggedIn;
-
-  return (
-    <div className={styles.errorContainer}>
-      <Message type="error" tagName="div">
-        <p>{getMessage()}</p>
-        {is403AndNotLoggedIn && (
-          <Button
-            type="link"
-            className={styles.userName}
-            onClick={() => {
-              const url = `${ROUTES.LOGIN}?to=${encodeURIComponent(
-                location.pathname
-              )}`;
-              navigate(url);
-            }}
-          >
-            Login
-          </Button>
-        )}
-      </Message>
-    </div>
-  );
-};
 
 interface MapProjectProps {
   /**
